Extract the setup orchestration policy into cli/integrations/
The four `nauro setup` commands and `nauro adopt` shared their
cross-surface wiring through `setup_all_surfaces` living in the setup
command module, while each single-surface command carried its own
orchestration inline. Moving the policy into
`cli/integrations/orchestrator.py` gives the command bodies a thin
flag-parse-then-echo shape and lets adopt depend on the integrations
layer instead of a sibling command module.

`setup_all_surfaces` moves verbatim (byte-identical). The single-surface
orchestration is extracted into `claude_code_surfaces`, `cursor_surfaces`,
and `codex_surfaces`, each returning the flat status lines the command
echoes. Claude Code's stderr warning routing (warn_then_regen's warn
callback) is threaded through an explicit `warn` parameter so the command
body still owns the `err=True` sink; the surface functions return plain
`list[str]` and change no observable stdout ordering.

The two notices shared by setup all and adopt move alongside the
orchestration; the three notices used only by the thin commands stay in
the command module.
