+++ packages/nauro/src/nauro/cli/commands/adopt.py
@@ -30,12 +30,12 @@ from pathlib import Path
 
 import typer
 
-from nauro.cli.commands.setup import (
+from nauro.cli.git_hygiene import public_surface_git_warnings
+from nauro.cli.integrations.orchestrator import (
     SHIP_TASK_NEEDS_SUBAGENTS_NOTICE,
     SUBAGENTS_CONNECTOR_NAME_NOTICE,
     setup_all_surfaces,
 )
-from nauro.cli.git_hygiene import public_surface_git_warnings
 from nauro.cli.integrations.skills import OPT_IN_SKILL_NAMES, SKILL_NAMES
 from nauro.cli.nauro_command import _find_nauro_command
 from nauro.cli.utils import refuse_global_config_collision, refuse_repo_config_symlink
